Beginnings of path groups
This patch introduces path groups, and also *loads* of bugs!

Prior to the patch the app had lists of paths at the root level. Since a real life object is usually constructed of multiple path cuts (ie: an outer shape and inner holes) this meant that you could only cut one shape at a time.

The restriction meant that you could not have multiple complex shapes cut in a single cutting session, which is inconvenient. Specifically it made importing SVG shapes awkward, which is the next step.

However loads of stuff is now broken and needs to be fixed.